requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
chardet==5.2.0
Django==5.0.2
django-admin-thumbnails==0.2.8
django-allauth==0.62.1
django-crispy-forms==2.1
django-grappelli==4.0.1
pillow==10.2.0
pycparser==2.22
reportlab==4.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
